python_ml/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
nltk
python-multipart
python-dotenv
openai
google-generativeai
